Correction d'un bug: on utilisait les coordonnées de l'AI et non celles du groupe pour déterminer la cible.

+++ src/IA/groupe.h
@@ -42,7 +42,7 @@ public:
 	const Case& cible() const;
 	bool pretAAttaquer();
 
-	int preparerAction();
+	double preparerAction();
 	void jouerAction();
 };
